requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.8.2
python-dotenv==1.0.1
apscheduler==3.10.4
requests==2.32.3